requirements.txt
aiogram
redis
python-dotenv
aioredis
APScheduler
requests

marshmallow